--- a/crypto/secp256k1/libsecp256k1/.gitattributes
+++ b/crypto/secp256k1/libsecp256k1/.gitattributes
@@ -1,0 +1,2 @@
+src/precomputed_ecmult.c linguist-generated
+src/precomputed_ecmult_gen.c linguist-generated