--- a/crypto/secp256k1/libsecp256k1/.gitignore
+++ b/crypto/secp256k1/libsecp256k1/.gitignore
@@ -1,17 +1,24 @@
-bench_inv
-bench_ecdh
-bench_sign
-bench_verify
-bench_schnorr_verify
-bench_recover
+bench
+bench_ecmult
 bench_internal
+noverify_tests
 tests
 exhaustive_tests
-gen_context
+precompute_ecmult_gen
+precompute_ecmult
+ctime_tests
+ecdh_example
+ecdsa_example
+schnorr_example
+ellswift_example
+musig_example
 *.exe
 *.so
 *.a
-!.gitignore
+*.csv
+*.log
+*.trs
+*.sage.py
 
 Makefile
 configure
@@ -21,6 +28,7 @@ aclocal.m4
 autom4te.cache/
 config.log
 config.status
+conftest*
 *.tar.gz
 *.la
 libtool
@@ -29,9 +37,15 @@ libtool
 *.lo
 *.o
 *~
-src/libsecp256k1-config.h
-src/libsecp256k1-config.h.in
-src/ecmult_static_context.h
+
+coverage/
+coverage.html
+coverage.*.html
+*.gcda
+*.gcno
+*.gcov
+
+build-aux/ar-lib
 build-aux/config.guess
 build-aux/config.sub
 build-aux/depcomp
@@ -45,5 +59,9 @@ build-aux/m4/ltversion.m4
 build-aux/missing
 build-aux/compile
 build-aux/test-driver
-src/stamp-h1
 libsecp256k1.pc
+
+### CMake
+/CMakeUserPresets.json
+# Default CMake build directory.
+/build