--- a/crypto/secp256k1/libsecp256k1/cmake/CheckArm32Assembly.cmake
+++ b/crypto/secp256k1/libsecp256k1/cmake/CheckArm32Assembly.cmake
@@ -1,0 +1,6 @@
+function(check_arm32_assembly)
+  try_compile(HAVE_ARM32_ASM
+    ${PROJECT_BINARY_DIR}/check_arm32_assembly
+    SOURCES ${PROJECT_SOURCE_DIR}/cmake/source_arm32.s
+  )
+endfunction()