--- a/crypto/secp256k1/libsecp256k1/cmake/config.cmake.in
+++ b/crypto/secp256k1/libsecp256k1/cmake/config.cmake.in
@@ -1,0 +1,5 @@
+@PACKAGE_INIT@
+
+include("${CMAKE_CURRENT_LIST_DIR}/@PROJECT_NAME@-targets.cmake")
+
+check_required_components(@PROJECT_NAME@)